[refactor] #199 - 뒷면 뷰 파일 분리 및 적용
