+++ Carmu/Sources/Views/SessionStart/SessionStartPassengerView.swift
@@ -23,7 +23,7 @@ final class SessionStartPassengerView: UIView {
     // 앞면 뷰
     lazy var passengerFrontView = PassengerFrontView()
     // 뒷면 뷰
-    private lazy var passengerBackView = PassengerBackView()
+    private lazy var sessionStartBackView = SessionStartBackView()
 
     init() {
         super.init(frame: .zero)
@@ -42,7 +42,7 @@ final class SessionStartPassengerView: UIView {
 
         contentView.frame = bounds
         passengerFrontView.frame = bounds
-        passengerBackView.frame = bounds
+        sessionStartBackView.frame = bounds
 
     }
 }
@@ -53,8 +53,8 @@ extension SessionStartPassengerView {
     private func setupUI() {
         addSubview(contentView)
         contentView.addSubview(passengerFrontView)
-        contentView.addSubview(passengerBackView)
-        passengerBackView.isHidden = true
+        contentView.addSubview(sessionStartBackView)
+        sessionStartBackView.isHidden = true
 
         layer.cornerRadius = 20
         layer.shadowColor = UIColor.semantic.accPrimary?.cgColor
@@ -73,7 +73,7 @@ extension SessionStartPassengerView {
 
         UIView.transition(with: contentView, duration: 0.4, options: transitionOptions, animations: {
             self.passengerFrontView.isHidden = self.isFlipped
-            self.passengerBackView.isHidden = !self.isFlipped
+            self.sessionStartBackView.isHidden = !self.isFlipped
         }, completion: nil)
     }
 }
@@ -243,56 +243,5 @@ final class PassengerFrontView: UIView {
         timeLabel.attributedText = timeText
 
         statusLabel.text = "운전자의 확인을 기다리고 있어요"
-    }
-}
-
-// MARK: - 뒷면 뷰
-
-final class PassengerBackView: UIView {
-
-    private lazy var personImage: UIImageView = {
-        let image = UIImage(systemName: "person.fill")
-        let imageView = UIImageView(image: image)
-        imageView.tintColor = UIColor.semantic.textPrimary
-        imageView.contentMode = .scaleAspectFit
-        return imageView
-    }()
-    private lazy var totalCrewMemeberLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont.carmuFont.subhead3
-        label.textColor = UIColor.semantic.textPrimary
-        return label
-    }()
-
-    init() {
-        super.init(frame: .zero)
-        setupBackView()
-        setupConstraints()
-        settingData()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private func setupBackView() {
-        addSubview(personImage)
-        addSubview(totalCrewMemeberLabel)
-    }
-
-    private func setupConstraints() {
-        personImage.snp.makeConstraints { make in
-            make.leading.top.equalToSuperview().inset(20)
-            make.width.height.equalTo(22)
-        }
-        totalCrewMemeberLabel.snp.makeConstraints { make in
-            make.leading.equalTo(personImage.snp.trailing).offset(4)
-            make.centerY.equalTo(personImage)
-        }
-    }
-
-    // TODO: - 실제 데이터로 변경
-    private func settingData() {
-        totalCrewMemeberLabel.text = "\(crewData?.crews.count ?? 0)명"
     }
 }
